pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "media-reviewer"
version = "0.1.0"
description = "A robust media procurement and curation tool."
requires-python = ">=3.10"
dependencies = [
    "flask~=3.1.2",
    "requests~=2.32.5",
    "python-dotenv~=1.2.1",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0"
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q"
